requirements.txt
RateMyProfessorAPI
sanic
motor
python-dotenv